move ThreadAsyncLock into hardware_control package.
